fix: Use absolute paths everywhere
Tanstack’s router is being inconsistent about relative routes. Prior to the upgrade, it was inconsistent with the `to` redirects. But now it’s behaving the same way with the `<Link />` components too. Maybe we’ll figure out why one day.

I made `/#/sign-in` for signing in, and `orgs` is now the actual top level route, removing `orgs` from the URL entirely. Both cloud and local mode should have consistent, handled routes now.

https://harperdb.atlassian.net/browse/STUDIO-395